+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `registerProtocolClasses` and `unregisterProtocolClasses`
 * Do not expose `records` on `VCRURLSessionCassette`, expose `numberOfRecords` instead
+* Add `numberOfPlayedRecords` to `VCRURLSessionCassette`
 
 ### 1.2.0 (2016-01-21)